backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
opencv-python==4.8.1.78
paddleocr==2.7.0.2
pillow==10.0.1
python-dotenv==1.0.0
openai==1.3.0
numpy==1.25.2
pydantic==2.4.2
pandas==2.1.1
pytesseract==0.3.10
pdf2image==1.16.3